+++ backend/packages/workspaces/CHANGELOG.md
@@ -1,5 +1,27 @@
 # @cat-factory/workspaces
 
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies [5c20968]
+  - @cat-factory/kernel@0.9.0
+
+## 0.7.4
+
+### Patch Changes
+
+- Updated dependencies [c70df09]
+  - @cat-factory/contracts@0.8.0
+  - @cat-factory/kernel@0.8.0
+
+## 0.7.3
+
+### Patch Changes
+
+- Updated dependencies [a0a1bcc]
+  - @cat-factory/kernel@0.7.3
+
 ## 0.7.2
 
 ### Patch Changes