+++ backend/runtimes/local/CHANGELOG.md
@@ -1,5 +1,39 @@
 # @cat-factory/local-server
 
+## 0.7.6
+
+### Patch Changes
+
+- Updated dependencies [5c20968]
+  - @cat-factory/kernel@0.9.0
+  - @cat-factory/server@0.8.0
+  - @cat-factory/agents@0.8.1
+  - @cat-factory/orchestration@0.7.5
+  - @cat-factory/node-server@0.7.5
+
+## 0.7.5
+
+### Patch Changes
+
+- Updated dependencies [c70df09]
+  - @cat-factory/agents@0.8.0
+  - @cat-factory/contracts@0.8.0
+  - @cat-factory/kernel@0.8.0
+  - @cat-factory/orchestration@0.7.4
+  - @cat-factory/server@0.7.4
+  - @cat-factory/node-server@0.7.4
+
+## 0.7.4
+
+### Patch Changes
+
+- Updated dependencies [a0a1bcc]
+  - @cat-factory/kernel@0.7.3
+  - @cat-factory/node-server@0.7.3
+  - @cat-factory/agents@0.7.3
+  - @cat-factory/orchestration@0.7.3
+  - @cat-factory/server@0.7.3
+
 ## 0.7.3
 
 ### Patch Changes